4.23

+++ part4/controllers/blogs.js
@@ -1,6 +1,7 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
+const middleware = require('../utils/middleware')
+
 
 blogsRouter.get('/', async (_request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
@@ -16,28 +17,38 @@ blogsRouter.get('/:id', async (request, response) => {
   }
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
   const body = request.body
-  const user = await User.findById(body.userId)
+  const user = request.user
 
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes | 0,
-    user: user._id,
+    user: user.id,
   })
 
   const savedBlog = await blog.save()
-  user.blogs = user.blogs.concat(savedBlog._id)
+  user.blogs = user.blogs.concat(savedBlog.id)
   await user.save()
 
-  response.json(savedBlog)
+  const populatedBlog = await savedBlog
+    .populate('user', { username: 1, name: 1 })
+
+  response.status(201).json(populatedBlog.toJSON())
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    return response.status(204).end()
+  } else {
+    return response.status(401).json({ error: 'missing permissions to delete blog' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
@@ -56,10 +67,11 @@ blogsRouter.put('/:id', async (request, response) => {
     context: 'query',
   })
 
-  if (updatedBlog)
-    response.json(updatedBlog)
-  else
+  if (updatedBlog) {
+    return response.json(updatedBlog)
+  } else {
     return response.status(404).end()
+  }
 })
 
 module.exports = blogsRouter
